refactor(controller): response -> range
